fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,15 +5,15 @@
  */
 int main(void)
 {
-	int firstNum = 1;
-	int secondNum = 2;
-	long long sum = firstNum + secondNum;
+	long firstNum = 1;
+	long secondNum = 2;
+	long sum = firstNum + secondNum;
 	int i;
 
-	printf("%d, %d, ", firstNum, secondNum);
+	printf("%ld, %ld, ", firstNum, secondNum);
 	for (i = 3; i < 50; i++)
 	{
-		printf("%lld", sum);
+		printf("%ld", sum);
 		firstNum = secondNum;
 		secondNum = sum;
 		sum = firstNum + secondNum;
